Put the date conversion methods in a static DateTools class
